little refactor

+++ functions.h
@@ -5,10 +5,14 @@
 #ifndef GBCLASS_FUNCTIONS_H
 #define GBCLASS_FUNCTIONS_H
 
-void showArray(const double* arr, int len);
-void mirrorArray(int* arr, int len);
-void inThree(int* arr, int len);
-void setArrayOffset(int* arr, int offset, int len);
-bool checkBalance(const int* arr, int len);
+#include <iostream>
+#include <algorithm>
+#include <numeric>
+
+void showArray(const double* arr, size_t len);
+void mirrorArray(int* arr, size_t len);
+void inThree(int* arr, size_t len);
+void setArrayOffset(int* arr, int offset, size_t len);
+bool checkBalance(const int* arr, size_t len);
 
 #endif  // GBCLASS_FUNCTIONS_H
